# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "## A Full-Featured EDTECH API built with FIBER \u0026 ENT ORM.\n\n\u003c!-- ### WEBSOCKETS:\n\n#### Notifications\n\n- URL: `wss://{host}/api/v4/ws/notifications`\n\n- Requires authorization, so pass in the Bearer Authorization header.\n\n- You can only read and not send notification messages into this socket. --\u003e\n\n\n\u003c!-- #### Chats\n\n- URL: `wss://{host}/api/v4/ws/chats/{id}`\n- Requires authorization, so pass in the Bearer Authorization header.\n- Use chat_id as the ID for an existing chat or username if it's the first message in a DM.\n- You cannot read realtime messages from a username that doesn't belong to the authorized user, but you can surely send messages.\n- Only send a message to the socket endpoint after the message has been created or updated, and files have been uploaded.\n- Fields when sending a message through the socket:\n\n  ```json\n  { \"status\": \"CREATED\", \"id\": \"fe4e0235-80fc-4c94-b15e-3da63226f8ab\" }\n  ``` --\u003e",
        "title": "EDNET API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/courses": {
            "get": {
                "description": "This endpoint retrieves paginated responses of latest courses",
                "tags": [
                    "Courses"
                ],
                "summary": "Retrieve Latest Courses",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Current Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Page Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter By Title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter By Instructor's Name Or Username",
                        "name": "instructor",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter By Free Status",
                        "name": "isFree",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort By Rating (asc or desc)",
                        "name": "sortByRating",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/courses.CoursesResponseSchema"
                        }
                    }
                }
            }
        },
        "/courses/lessons/{slug}/quizzes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This endpoint retrieves paginated responses of a lesson quizzes",
                "tags": [
                    "Courses"
                ],
                "summary": "Retrieve Lesson Quizzes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lesson Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Current Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Page Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter By Title",
                        "name": "title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/courses.QuizzesResponseSchema"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/base.NotFoundErrorExample"
                        }
                    }
                }
            }
        },
        "/courses/pdf/summarize": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint accepts a PDF file and returns a summarized version of its content.`",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Courses"
                ],
                "summary": "Summarize a PDF document",
                "parameters": [
                    {
                        "type": "file",
                        "description": "PDF file to summarize",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "Maximum number of summary points",
                        "name": "max_points",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/courses.PDFSummaryResponseSchema"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/base.InvalidErrorExample"
                        }
                    }
                }
            }
        },
        "/courses/quizzes/{quiz_slug}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This endpoint retrieves the details of a particular quiz",
                "tags": [
                    "Courses"
                ],
                "summary": "Retrieve Quiz Details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quiz Slug",
                        "name": "quiz_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/courses.QuizResponseSchema"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/base.NotFoundErrorExample"
                        }
                    }
                }
            }
        },
        "/courses/quizzes/{quiz_slug}/results": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint retrieves the result of a particular quiz for a user`",
                "tags": [
                    "Courses"
                ],
                "summary": "Retrieve Quiz Result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quiz Slug",
                        "name": "quiz_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/courses.QuizResultResponseSchema"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/base.NotFoundErrorExample"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint allows a user to submit their answers for a quiz`\n`If this submission is for the last quiz in the course, a certificate will be generated`",
                "tags": [
                    "Courses"
                ],
                "summary": "Submit Quiz",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quiz Slug",
                        "name": "quiz_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Submission object",
                        "name": "result",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/courses.QuizSubmissionSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/courses.QuizResultResponseSchema"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/base.InvalidErrorExample"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/base.NotFoundErrorExample"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/base.ValidationErrorExample"
                        }
                    }
                }
            }
        },
        "/courses/quizzes/{quiz_slug}/start": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "`This endpoint allows a user to start a quiz`",
                "tags": [
                    "Courses"
                ],
                "summary": "Start Quiz",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quiz Slug",
                        "name": "quiz_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/base.ResponseSchema"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/base.InvalidErrorExample"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/base.NotFoundErrorExample"
                        }
                    }
                }
            }
        },
        "/courses/{course_slug}/lessons/{lesson_slug}": {
            "get": {
                "description": "This endpoint retrieves the details of a particular lesson",
                "tags": [
                    "Courses"
                ],
                "summary": "Retrieve Lesson Details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course Slug",
                        "name": "course_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Lesson Slug",
                        "name": "lesson_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/courses.LessonResponseSchema"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/base.NotFoundErrorExample"
                        }
                    }
                }
            }
        },
        "/courses/{slug}": {
            "get": {
                "description": "This endpoint retrieves the details of a particular course",
                "tags": [
                    "Courses"
                ],
                "summary": "Retrieve Course Details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/courses.CourseResponseSchema"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/base.NotFoundErrorExample"
                        }
                    }
                }
            }
        },
        "/courses/{slug}/enroll": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "This endpoint allows a user to enroll for a specific course",
                "tags": [
                    "Courses"
                ],
                "summary": "Enroll for a course",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Enrollment object",
                        "name": "enrollment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/courses.EnrollForACourseSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/courses.EnrollmentResponseSchema"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/base.InvalidErrorExample"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/base.NotFoundErrorExample"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/base.ValidationErrorExample"
                        }
                    }
                }
            }
        },
        "/courses/{slug}/lessons": {
            "get": {
                "description": "This endpoint retrieves paginated responses of a course lessons",
                "tags": [
                    "Courses"
                ],
                "summary": "Retrieve Course Lessons",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Course Slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Current Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Page Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter By Title",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter By Free Preview",
                        "name": "isFreePreview",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/courses.LessonsResponseSchema"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/base.NotFoundErrorExample"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "courses.CoursesResponseSchema": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/config.PaginationResponse-courses_CourseListSchema"
                },
                "message": {
                    "type": "string",
                    "example": "Data fetched/created/updated/deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "courses.QuizzesResponseSchema": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/config.PaginationResponse-courses_QuizListSchema"
                },
                "message": {
                    "type": "string",
                    "example": "Data fetched/created/updated/deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer jwt_string' to correctly set the API Key",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}